requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
numpy<2
gunicorn==21.2.0
python-dotenv==1.0.0
